Move onPublicStageDataUpdated trigger for chat timer logic to separate file
